renamed packaags and links as needed
