+++ python/MANIFEST
@@ -1,6 +1,6 @@
 # file GENERATED by distutils, do NOT edit
 setup.cfg
 setup.py
-simplesauce/__init__.py
-simplesauce/options.py
-simplesauce/session.py
+saucebindings/__init__.py
+saucebindings/options.py
+saucebindings/session.py
